@@ -1,49 +1,30 @@
+import {
+  Observable
+} from "rxjs";
 import { Injectable } from "@angular/core";
-import { Observable } from "rxjs/Observable";
-import { StompHeaders } from "@stomp/stompjs";
-import * as SockJS from "sockjs-client";
-import * as Stomp from "@stomp/stompjs";
+import { StompService } from "@stomp/ng2-stompjs";
+import { map } from "rxjs/operators";
+import { Message } from "@stomp/stompjs";
+
 declare let sessionStorage: Storage;
 
 @Injectable()
 export class StompProvider {
-  constructor() {}
-
-  socketProvider() {
-    return new SockJS("http://yummy-backend.herokuapp.com/api/topics/panel");
-  }
+  constructor(private stompService: StompService) {}
 
   public connectSockets<T>(endpoint: string): Observable<T> {
     return Observable.create((observer) => {
-      let socket = this.socketProvider();
-      let stompClient = Stomp.over(socket);
-      stompClient.connect(
-        {
-          "Authorization": sessionStorage.__th,
-          "Email": sessionStorage.__mail
-        },
-        (frame) => {
-          console.log("Connected: " + frame);
-          stompClient.subscribe("/topic/orders/current", (greeting) => {
-            observer.next(greeting)
-          });
-        }
-      );
+      let sub = this.stompService.subscribe(endpoint, {
+        login: "guest",
+        passcode: "guest"
+      });
+
+      sub
+        .pipe(map((message: Message) => message.body))
+        .subscribe((msg_body: string) => {
+          //console.log(`Received: ${msg_body}`);
+          observer.next(msg_body);
+        });
     });
-  }
-
-  private getHeadersValues(): Observable<any[]> {
-    return Observable.combineLatest(
-      Observable.of(sessionStorage.__th),
-      Observable.of(sessionStorage.__mail)
-    );
-  }
-
-  private getHeaders(authToken, login): StompHeaders {
-    let headers = new StompHeaders();
-    headers.authToken = authToken;
-    headers.login = login;
-
-    return headers;
   }
 }
