@@ -6,13 +6,15 @@ import { AppState } from "../../core/app-state";
 import * as fromModalAction from '../../core/actions/_modal.actions';
 import { Store } from "@ngrx/store";
 import { _ModalType } from "../../core/models/_modal";
+import { format } from "date-fns";
 
 @Component({
   selector: "offers-container",
   templateUrl: "offers-container.html"
 })
 export class OffersContainerComponent {
-  offers$: Observable<Offer[]>;
+  public offers$: Observable<Offer[]>;
+  public date: string = format(new Date(), 'DD-MM-YYYY');
 
   constructor(private store: Store<AppState>) {}
 
@@ -20,7 +22,7 @@ export class OffersContainerComponent {
     this.offers$ = this.store.select((state) => state.offer.data);
   }
 
-  add(): void {
+  public add(): void {
     this.store.dispatch(new fromModalAction.Show({
       mode: _ModalType.OFFER,
       meta: {
