@@ -3,6 +3,7 @@ import { Offer } from '../../../core/models/offer';
 import { Observable } from 'rxjs';
 import { Store } from '@ngrx/store';
 import { AppState } from '../../../core/app-state';
+import * as fromOfferActions from "../../../core/actions/offer.actions";
 
 @Component({
   selector: 'offers-list',
@@ -10,15 +11,20 @@ import { AppState } from '../../../core/app-state';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class OffersListComponent {
-
-  offers: Observable<Offer[]>;
-  isFetching: Observable<boolean>;
+  public page$: Observable<number>;
+  public offers$: Observable<Offer[]>;
+  public isFetching$: Observable<boolean>;
 
   constructor(private store: Store<AppState>) {}
 
   ngOnInit(): void {
-    this.offers = this.store.select((state) => state.offer.data);
-    this.isFetching = this.store.select((state) => state.offer.isFetching);
+    this.page$ = this.store.select((state) => state.offer.page);
+    this.offers$ = this.store.select((state) => state.offer.data);
+    this.isFetching$ = this.store.select((state) => state.offer.isFetching);
+  }
+
+  public paginate(page: number): void {
+    this.store.dispatch(new fromOfferActions.PaginateOffers(page))
   }
 
 }
