@@ -3,6 +3,7 @@ import { Pending } from "../../../core/models/pending";
 import { Observable } from "rxjs";
 import { Store } from "@ngrx/store";
 import { AppState } from "../../../core/app-state";
+import * as fromPendingActions from "../../../core/actions/pending.actions";
 
 @Component({
   selector: "pending-list",
@@ -10,13 +11,19 @@ import { AppState } from "../../../core/app-state";
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class PendingListComponent {
-  pendings$: Observable<Pending[]>;
-  isFetching$: Observable<boolean>;
+  public pendings$: Observable<Pending[]>;
+  public isFetching$: Observable<boolean>;
+  public page$: Observable<number>;
 
   constructor(private store: Store<AppState>) {}
 
   ngOnInit(): void {
     this.pendings$ = this.store.select((state) => state.pending.data);
     this.isFetching$ = this.store.select((state) => state.pending.isFetching);
+    this.page$ = this.store.select((state) => state.pending.page);
+  }
+
+  public paginate(page: number): void {
+    this.store.dispatch(new fromPendingActions.PaginatePendings(page))
   }
 }
