@@ -3,6 +3,7 @@ import { Action } from "@ngrx/store";
 export const FETCH_TRANSACTIONS = "[Transaction] Fetch Transactions";
 export const FETCH_TRANSACTIONS_SUCC = "[Transaction] Fetch Transactions Succ";
 export const FETCH_TRANSACTIONS_FAIL = "[Transaction] Fetch Transactions Fail";
+export const PAGINATE_TRANSACTIONS = "[Transaction] Paginate Transactions";
 
 export class FetchTransactions implements Action {
   readonly type = FETCH_TRANSACTIONS;
@@ -18,7 +19,13 @@ export class FetchTransactionsFail implements Action {
   readonly type = FETCH_TRANSACTIONS_FAIL;
 }
 
+export class PaginateTransactions implements Action {
+  readonly type = PAGINATE_TRANSACTIONS;
+  constructor(public payload: number) {}
+}
+
 export type TransactionActions =
   | FetchTransactions
   | FetchTransactionsSucc
-  | FetchTransactionsFail;
+  | FetchTransactionsFail
+  | PaginateTransactions;
