@@ -1,3 +1,5 @@
+
+import {map, pluck, tap, switchMap} from 'rxjs/operators';
 import { Injectable } from "@angular/core";
 import { Actions, Effect } from "@ngrx/effects";
 import * as fromActions from "../actions/offer.actions";
@@ -18,48 +20,38 @@ export class OfferEffects {
     private stompProvider: StompProvider
   ) {}
 
-  // @Effect()
-  // public getOffers$ = this.actions$
-  //   .ofType(fromActions.FETCH_OFFERS)
-  //   .switchMap(() => this.offerProvider.getOffers())
-  //   .map((res: any | HttpErrorResponse) => {
-  //     return (!(res instanceof HttpErrorResponse) && res.offers)
-  //       ? new fromActions.FetchOffersSucc(res.offers)
-  //       : new fromActions.FetchOffersFail;
-  //   });
-
   @Effect()
   public deleteOffer$ = this.actions$
-    .ofType(fromActions.DELETE_OFFER)
-    .pluck("payload")
-    .do(() => this.store.dispatch(new fromLoaderActions.Show("Usuwanie oferty...")))
-    .switchMap((payload: Offer) => this.offerProvider.deleteOffer(payload))
-    .do(() => this.store.dispatch(new fromLoaderActions.Hide))
-    .map((res: Offer | HttpErrorResponse) => {
+    .ofType(fromActions.DELETE_OFFER).pipe(
+    pluck("payload"),
+    tap(() => this.store.dispatch(new fromLoaderActions.Show("Usuwanie oferty..."))),
+    switchMap((payload: Offer) => this.offerProvider.deleteOffer(payload)),
+    tap(() => this.store.dispatch(new fromLoaderActions.Hide)),
+    map((res: Offer | HttpErrorResponse) => {
       return (!(res instanceof HttpErrorResponse))
         ? new fromActions.DeleteOfferSucc(res)
         : new fromActions.DeleteOfferFail;
-    });
+    }),);
 
   @Effect()
   public updateOffer$ = this.actions$
-    .ofType(fromActions.UPDATE_OFFER)
-    .pluck("payload")
-    .do(() => this.store.dispatch(new fromLoaderActions.Show("Aktualizacja oferty...")))
-    .switchMap((payload: Offer) => this.offerProvider.updateOffer(payload))
-    .do(() => this.store.dispatch(new fromLoaderActions.Hide))
-    .map((res: Offer | HttpErrorResponse) => {
+    .ofType(fromActions.UPDATE_OFFER).pipe(
+    pluck("payload"),
+    tap(() => this.store.dispatch(new fromLoaderActions.Show("Aktualizacja oferty..."))),
+    switchMap((payload: Offer) => this.offerProvider.updateOffer(payload)),
+    tap(() => this.store.dispatch(new fromLoaderActions.Hide)),
+    map((res: Offer | HttpErrorResponse) => {
       return (!(res instanceof HttpErrorResponse))
         ? new fromActions.UpdateOfferSucc(res)
         : new fromActions.UpdateOfferFail;
-    });
+    }),);
 
   @Effect()
   public socketOffers$ = this.stompProvider
-    .connectSockets('/topic/orders/current')
-    .map((res: any | HttpErrorResponse) => {
+    .connectSockets('/topic/orders/current').pipe(
+    map((res: any | HttpErrorResponse) => {
       return (!(res instanceof HttpErrorResponse) && res.offers)
         ? new fromActions.FetchOffersSucc(res.offers)
         : new fromActions.FetchOffersFail;
-    });
+    }));
 }
