@@ -1,3 +1,5 @@
+
+import {map, switchMap} from 'rxjs/operators';
 import { Actions, Effect } from "@ngrx/effects";
 import { Injectable } from "@angular/core";
 import * as fromActions from "../actions/transaction.actions";
@@ -16,12 +18,12 @@ export class TransactionEffects {
 
   @Effect()
   public getPendings$ = this.actions$
-    .ofType(fromActions.FETCH_TRANSACTIONS)
-    .switchMap(() => this.transactionProvider.getTransactions())
-    .map((res: any | HttpErrorResponse) => {
+    .ofType(fromActions.FETCH_TRANSACTIONS).pipe(
+    switchMap(() => this.transactionProvider.getTransactions()),
+    map((res: any | HttpErrorResponse) => {
       return (!(res instanceof HttpErrorResponse) && res.currentOrderList)
         ? new fromActions.FetchTransactionsSucc(res.currentOrderList)
         : new fromActions.FetchTransactionsFail();
-    });
+    }),);
 
 }
