@@ -1,9 +1,11 @@
+
+import {of as observableOf,  Observable } from 'rxjs';
+
+import {mergeMap, tap, catchError} from 'rxjs/operators';
 import { Injectable } from "@angular/core";
 import { Rest } from "../rest/rest";
 import { Settings } from "../settings/settings";
 import { LoginForm } from "../../models/login-form";
-
-import { Observable } from "rxjs";
 import { LoginUserBody } from "../rest/rest";
 import { HttpErrorResponse } from "@angular/common/http";
 
@@ -17,14 +19,14 @@ export class LoginProvider {
   ) {}
 
   loginUser(loginForm: LoginForm): Observable<string | HttpErrorResponse> {
-    return this.rest.loginUser(this.parseForm(loginForm))
-      .mergeMap((data) => this.saveAuth(data))
-      .catch((err) => Observable.of(err));
+    return this.rest.loginUser(this.parseForm(loginForm)).pipe(
+      mergeMap((data) => this.saveAuth(data)),
+      catchError((err) => observableOf(err)),);
   }
 
   saveAuth({ Authorization }): Observable<void> {
-    return Observable.of(null)
-      .do(() => sessionStorage.__th = Authorization);
+    return observableOf(null).pipe(
+      tap(() => sessionStorage.__th = Authorization));
   }
 
   parseForm(loginForm: LoginForm): LoginUserBody {
